chore: rename project from weathercloset to classify and update dependencies
