+++ android/app/src/main/kotlin/com/example/classify/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.weathercloset
+package com.example.classify
 
 import io.flutter.embedding.android.FlutterActivity
