+++ lib/data/services/firebase_auth_service.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:weathercloset/global/global.dart';
+import 'package:classify/global/global.dart';
 import 'package:flutter/foundation.dart';
-import 'package:weathercloset/data/services/google_login_service.dart';
+import 'package:classify/data/services/google_login_service.dart';
 class FirebaseAuthService {
   final FirebaseAuth _auth = firebaseAuth;
