+++ lib/ui/archive/archive_view/widgets/buildStudyDetailPage.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:weathercloset/domain/models/memo/memo_model.dart';
+import 'package:classify/domain/models/memo/memo_model.dart';
 import 'package:intl/intl.dart';
-import 'package:weathercloset/ui/archive/archive_view/view_models/archive_view_model.dart';
-import 'package:weathercloset/utils/top_level_setting.dart';
+import 'package:classify/ui/archive/archive_view/view_models/archive_view_model.dart';
+import 'package:classify/utils/top_level_setting.dart';
 
 class StudyDetailPage extends StatefulWidget {
   final MemoModel memo;
