+++ lib/ui/archive/archive_view/widgets/buildTodoTabview.dart
@@ -1,7 +1,7 @@
-import 'package:weathercloset/domain/models/memo/memo_model.dart';
+import 'package:classify/domain/models/memo/memo_model.dart';
 import 'package:flutter/material.dart';
-import 'package:weathercloset/ui/archive/archive_view/view_models/archive_view_model.dart';
-import 'package:weathercloset/utils/top_level_setting.dart';
+import 'package:classify/ui/archive/archive_view/view_models/archive_view_model.dart';
+import 'package:classify/utils/top_level_setting.dart';
 
 Widget buildTodoTabView(Map<String, MemoModel> memos, ArchiveViewModel viewModel) {
   // '할 일' 카테고리만 필터링
