+++ lib/ui/today_act/view_models/today_act_view_model.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:weathercloset/data/repositories/memo/memo_repository.dart';
-import 'package:weathercloset/domain/models/memo/memo_model.dart';
+import 'package:classify/data/repositories/memo/memo_repository.dart';
+import 'package:classify/domain/models/memo/memo_model.dart';
 
 class TodayActViewModel extends ChangeNotifier {
   final MemoRepository _memoRepository;
