+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: weathercloset
+name: classify
 description: "A new Flutter project."
 publish_to: 'none'
 version: 0.1.0
@@ -34,6 +34,7 @@ dependencies:
   intl: ^0.19.0
   cloud_functions: ^5.4.0
   google_sign_in: ^6.3.0
+  package_rename: ^1.10.0
 
 dev_dependencies:
   flutter_test:
@@ -45,3 +46,9 @@ flutter:
   uses-material-design: true
   assets:
     - assets/woman_model.jpg
+
+package_rename_config:
+  android:
+    app_name: Classify
+    package_name: com.example.classify
+    override_old_package: com.example.weathercloset
